Use a Renderer interface instead of rendering fragments
